requirements.txt
flask
flask_sqlalchemy
flask_migrate
boto3
requests
uuid
python-dotenv
pymysql
flask-cors

